requirements.txt
flask
tensorflow
opencv-python
pillow
numpy
torch
torchvision
